add components in informationBlock
