+++ frontend/src/components/informationBlockComponents/SmallItem.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const SmallItem = ({ text }) => (
+    <div className="small-item">
+        {text}
+    </div>
+);
+
+export default SmallItem;
